+++ lib/decoding/decoders/index.rb
@@ -23,7 +23,7 @@ module Decoding
       # @param value [Object]
       # @return [Decoding::Decoder<Object>]
       def call(value)
-        return err(failure("expected an Array, got: #{value.class}")) unless value.is_a?(::Array)
+        return err(failure("expected Array, got #{value.class}")) unless value.is_a?(::Array)
 
         @decoder
           .call(value.fetch(@index))