+++ spec/decoding/decoders/match_spec.rb
@@ -11,9 +11,9 @@ module Decoding
       end
 
       it "errors with values not matching the given pattern" do
-        expect(Decoding.decode(Match.new(/foo/), 123)).to eql(Result.err("expected value matching /foo/, got: 123"))
+        expect(Decoding.decode(Match.new(/foo/), 123)).to eql(Result.err("expected value matching /foo/, got 123"))
         expect(Decoding.decode(Match.new(String), nil)).to eql(Result.err("expected String, got NilClass"))
-        expect(Decoding.decode(Match.new("other"), nil)).to eql(Result.err("expected value matching \"other\", got: nil"))
+        expect(Decoding.decode(Match.new("other"), nil)).to eql(Result.err("expected value matching \"other\", got nil"))
       end
     end
   end